+++ custom_components/blueiris/helpers/advanced_configurations_generator.py
@@ -1,9 +1,11 @@
 import logging
+from typing import List
 
 from homeassistant.components.media_player import SUPPORT_PLAY_MEDIA
 from homeassistant.core import HomeAssistant
 from homeassistant.util import slugify
 
+from ..models.camera_data import CameraData
 from ..models.entity_data import EntityData
 from .const import *
 
@@ -52,21 +54,21 @@ class AdvancedConfigurationGenerator:
             _add_to_file(out, ui_lovelace)
 
     @staticmethod
-    def generate_input_select_camera(camera_list):
+    def generate_input_select_camera(camera_list: List[CameraData]):
         entity_name = "BlueIris Camera"
         camera_items = []
         initial_camera = None
-        for camera_details in camera_list:
-            camera_id = camera_details.get("optionDisplay")
-            camera_name = camera_details.get("optionDisplay")
 
-            if camera_id in SYSTEM_CAMERA_ID:
-                initial_camera = camera_name
+        for camera in camera_list:
+            if camera.is_system:
+                initial_camera = camera.name
 
-            camera_items.append(camera_name)
+            camera_items.append(camera.name)
 
         if initial_camera is None:
-            initial_camera = camera_items[0]
+            for name in camera_items:
+                initial_camera = name
+                break
 
         lines = [
             f"  {slugify(entity_name)}:",
@@ -223,7 +225,7 @@ class AdvancedConfigurationGenerator:
             else:
                 lines.append("        entities: []")
 
-    def generate_script(self, camera_list, media_players):
+    def generate_script(self, camera_list: List[CameraData], media_players):
         entity_name = "BlueIris Cast"
         lines = [
             f"  {slugify(entity_name)}:",
@@ -259,10 +261,7 @@ class AdvancedConfigurationGenerator:
 
         camera_entity_ids = []
         for camera in camera_list:
-            camera_id = camera.get("optionValue")
-            camera_name = camera.get("optionDisplay")
-
-            camera_entity_ids.append(f"'{camera_name}': '{camera_id}'")
+            camera_entity_ids.append(f"'{camera.name}': '{camera.id}'")
 
         camera_entities = ", ".join(camera_entity_ids)
         lines.append(f"            {{% set camera_list = {{{camera_entities}}} %}}")