+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - Removed NONE option from drop-down, NONE was workaround for a validation issue in Integration's Options and fixed as part of HA v0.112.0
 - Reduced code of camera's configuration in Integration's Options 
 
+**Fixed bugs:**
+- Fixed - Generate configuration files
+
 ## 2020-07-17
 
 **Implemented enhancements:**